Extract settings logic into module
