+++ js/game.js
@@ -1,6 +1,7 @@
 import { controls, initControls } from "./controls.js";
 import { DayNightCycle } from './dayNight.js';
 import { Rabbit } from './rabbit.js';
+import { initSettings } from './settings.js';
 import * as THREE from 'https://unpkg.com/three@0.158.0/build/three.module.js';
 export function startGame() {
 
@@ -11,53 +12,6 @@ renderer.setPixelRatio(Math.min(devicePixelRatio, 2));
 renderer.setSize(innerWidth, innerHeight);
 renderer.shadowMap.enabled = true;
 app.appendChild(renderer.domElement);
-
-// UI controls
-const ballSlider = document.getElementById('ballSlider');
-const ballCountLabel = document.getElementById('ballCountLabel');
-const settings = document.getElementById('settings');
-const volumeSlider = document.getElementById('volumeSlider');
-const volumeLabel = document.getElementById('volumeLabel');
-const faceSlider = document.getElementById('faceSlider');
-const faceLabel = document.getElementById('faceLabel');
-const rabbitHealthSlider = document.getElementById('rabbitHealthSlider');
-const rabbitHealthLabel = document.getElementById('rabbitHealthLabel');
-const bulletDamageSlider = document.getElementById('bulletDamageSlider');
-const bulletDamageLabel = document.getElementById('bulletDamageLabel');
-const ballDamageSlider = document.getElementById('ballDamageSlider');
-const ballDamageLabel = document.getElementById('ballDamageLabel');
-
-let maxBalls = parseInt(ballSlider.value);
-ballSlider.addEventListener('input', () => {
-  ballCountLabel.textContent = ballSlider.value;
-  maxBalls = parseInt(ballSlider.value);
-});
-
-let bulletDamage = parseInt(bulletDamageSlider.value);
-bulletDamageSlider.addEventListener('input', () => {
-  bulletDamage = parseInt(bulletDamageSlider.value);
-  bulletDamageLabel.textContent = bulletDamage;
-});
-
-let ballDamage = parseInt(ballDamageSlider.value);
-ballDamageSlider.addEventListener('input', () => {
-  ballDamage = parseInt(ballDamageSlider.value);
-  ballDamageLabel.textContent = ballDamage;
-});
-
-Rabbit.faceOffset = parseFloat(faceSlider.value);
-faceSlider.addEventListener('input', () => {
-  const v = parseFloat(faceSlider.value);
-  faceLabel.textContent = v.toFixed(2);
-  Rabbit.faceOffset = v;
-});
-
-let rabbits = [];
-rabbitHealthSlider.addEventListener('input', () => {
-  const v = parseInt(rabbitHealthSlider.value);
-  rabbitHealthLabel.textContent = v;
-  for (const r of rabbits) { r.maxHealth = v; r.health = v; }
-});
 
 const scene = new THREE.Scene();
 scene.background = new THREE.Color(0x7fb0ff);
@@ -76,12 +30,10 @@ scene.fog = new THREE.Fog(0x7fb0ff, 20, 140);
     nightSound.setLoop(true);
     nightSound.setVolume(1.0);
   });
-  listener.setMasterVolume(parseFloat(volumeSlider.value));
-  volumeSlider.addEventListener('input', () => {
-    const v = parseFloat(volumeSlider.value);
-    volumeLabel.textContent = v.toFixed(2);
-    listener.setMasterVolume(v);
-  });
+
+let rabbits = [];
+const { settings, onPointerLockChange, getMaxBalls, getBulletDamage, getBallDamage } =
+  initSettings(renderer.domElement, rabbits, listener);
 
 // --- Lights ---
 const hemi = new THREE.HemisphereLight(0xffffff, 0x335533, 0.6);
@@ -280,10 +232,7 @@ function handleClick(){
     shootBullet();
   }
 }
-function onPointerLockChange(locked){
-  settings.classList.toggle('hidden', locked);
-}
-initControls(renderer.domElement, handleClick, onPointerLockChange);
+initControls(renderer.domElement, handleClick, onPointerLockChange, () => settings.classList.contains('hidden'));
 
 
 
@@ -339,7 +288,7 @@ const GRAV = 22;
 
   // Spawn balls from dispensers
   for (const d of dispensers) {
-    if (totalDispensed >= maxBalls) break;
+    if (totalDispensed >= getMaxBalls()) break;
     if (now - d.last >= 1000) {
       const mesh = new THREE.Mesh(ballGeo, ballMat);
       mesh.position.copy(d.pos).add(new THREE.Vector3(0, BALL_RADIUS, 0));
@@ -369,7 +318,7 @@ const GRAV = 22;
       if (!r.visible) continue;
       const radius = BALL_RADIUS * b.mesh.scale.x;
       if (r.mesh.position.distanceTo(b.mesh.position) < radius + 1) {
-        r.hitByBall(ballDamage);
+        r.hitByBall(getBallDamage());
         scene.remove(b.mesh); balls.splice(i,1);
         break;
       }
@@ -401,7 +350,7 @@ const GRAV = 22;
     for (const r of rabbits) {
       if (!r.visible) continue;
       if (b.mesh.position.distanceTo(r.mesh.position) < 2) {
-        r.damage(bulletDamage);
+        r.damage(getBulletDamage());
         scene.remove(b.mesh); bullets.splice(i,1);
         break;
       }
